make this compile with -DDEBUG_SVR4 [lossage because sizeof is u_long now]

+++ sys/compat/svr4/svr4_sockio.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: svr4_sockio.c,v 1.12 1999/09/07 18:20:19 christos Exp $	 */
+/*	$NetBSD: svr4_sockio.c,v 1.13 2000/04/12 15:12:13 christos Exp $	 */
 
 /*-
  * Copyright (c) 1995 The NetBSD Foundation, Inc.
@@ -165,8 +165,9 @@ svr4_sock_ioctl(fp, p, retval, fd, cmd, data)
 			if ((error = copyin(data, &sc, sizeof(sc))) != 0)
 				return error;
 
-			DPRINTF(("ifreq %d svr4_ifreq %d ifc_len %d\n",
-				sizeof(struct ifreq), sizeof(struct svr4_ifreq),
+			DPRINTF(("ifreq %ld svr4_ifreq %ld ifc_len %d\n",
+				(unsigned long)sizeof(struct ifreq),
+				(unsigned long)sizeof(struct svr4_ifreq),
 				sc.svr4_ifc_len));
 
 			if ((error = (*ctl)(fp, OSIOCGIFCONF,
